Remove solr3
Because nobody else supports it either and it is impossible
to make changes that are guaranteed to work with solr3.
Solr 3 uses a tomcat for which you cannot find documentation online any longer.
